environment: move compression level into repo settings
The compression level for loose objects and packfiles is tracked via a
couple of global variables. Refactor these so that the values for them
are tracked via repo settings, which allows us to drop this global
dependency.

Note that the refactoring is mostly straight-forward, except in
git-pack-objects(1). Here it is possible to change the compression level
via a command line option, and that option of course should override
whatever the user has configured. This creates the problem that we need
to be able to see whether the option has been given in the first place.

This is done by using `INT_MIN` as default value. Any value smaller than
-1 is an invalid compression level, so it's quite unlikely that any user
ever passed that sentinel value. And if they did we would have died
anyway.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>